300 - Move connect four logic from controller to model
