+++ src/Game/Games/ConnectFour/ConnectFourScene.java
@@ -7,13 +7,11 @@ import Game.Games.ConnectFour.ConnectFourView.ConnectFourView;
 import Player.Player;
 
 public class ConnectFourScene extends GameScene {
-    private final static int DEFAULT_ROW = 6;
-    private final static int DEFAULT_COLUMN = 7;
 
     public ConnectFourScene(Player[] listPlayers, boolean isTraining, int[] scores) {
-        this.model = new ConnectFourModel(listPlayers, DEFAULT_ROW, DEFAULT_COLUMN);
-        this.view = new ConnectFourView(DEFAULT_ROW, DEFAULT_COLUMN, listPlayers, scores);
-        this.controller = new ConnectFourController((ConnectFourModel) this.model, (ConnectFourView) this.view, DEFAULT_ROW, DEFAULT_COLUMN, isTraining);
+        this.model = new ConnectFourModel(listPlayers);
+        this.view = new ConnectFourView(listPlayers, scores);
+        this.controller = new ConnectFourController((ConnectFourModel) this.model, (ConnectFourView) this.view, isTraining);
         this.controller.addObserver(this);
     }
 }
